[Rename] 좋아요버튼 png파일로 변경 후 다시 연결 #52
